+++ mobile/android/android-components/components/support/ktx/src/test/java/mozilla/components/support/ktx/android/view/ViewTest.kt
@@ -4,16 +4,27 @@
 
 package mozilla.components.support.ktx.android.view
 
+import android.app.Activity
 import android.view.View
+import android.view.WindowManager
 import android.widget.EditText
 import android.widget.TextView
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.asCoroutineDispatcher
+import kotlinx.coroutines.isActive
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.test.resetMain
+import kotlinx.coroutines.test.setMain
 import mozilla.components.support.base.android.Padding
 import mozilla.components.support.test.any
 import mozilla.components.support.test.robolectric.testContext
+import org.junit.After
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertTrue
+import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.mockito.Mockito.`when`
@@ -22,10 +33,27 @@ import org.mockito.Mockito.doReturn
 import org.mockito.Mockito.never
 import org.mockito.Mockito.spy
 import org.mockito.Mockito.verify
+import org.robolectric.Robolectric
 import org.robolectric.shadows.ShadowLooper
+import java.util.concurrent.CountDownLatch
+import java.util.concurrent.Executors
+import java.util.concurrent.TimeUnit
 
 @RunWith(AndroidJUnit4::class)
 class ViewTest {
+    @Before
+    @ExperimentalCoroutinesApi
+    fun setUp() {
+        // We create a separate thread for the main dispatcher so that we do not deadlock our test
+        // thread.
+        Dispatchers.setMain(Executors.newSingleThreadExecutor().asCoroutineDispatcher())
+    }
+
+    @After
+    @ExperimentalCoroutinesApi
+    fun tearDown() {
+        Dispatchers.resetMain()
+    }
 
     @Test
     fun `showKeyboard should request focus`() {
@@ -129,5 +157,60 @@ class ViewTest {
         viewTreeObserver.dispatchOnGlobalLayout()
 
         verify(viewTreeObserver).removeOnGlobalLayoutListener(any())
+    }
+
+    @Test(expected = IllegalStateException::class)
+    fun `toScope throws if view is not attached`() {
+        val view = View(testContext)
+        view.toScope()
+    }
+
+    @Test
+    fun `can dispatch coroutines to view scope`() {
+        val activity = Robolectric.buildActivity(Activity::class.java).create().get()
+        val view = View(testContext)
+        activity.windowManager.addView(view, WindowManager.LayoutParams(100, 100))
+
+        assertTrue(view.isAttachedToWindow)
+
+        val latch = CountDownLatch(1)
+        var coroutineExecuted = false
+
+        view.toScope().launch {
+            coroutineExecuted = true
+            latch.countDown()
+        }
+
+        latch.await(10, TimeUnit.SECONDS)
+
+        assertTrue(coroutineExecuted)
+    }
+
+    @Test
+    fun `scope is cancelled when view is detached`() {
+        val activity = Robolectric.buildActivity(Activity::class.java).create().get()
+        val view = View(testContext)
+        activity.windowManager.addView(view, WindowManager.LayoutParams(100, 100))
+
+        val scope = view.toScope()
+
+        assertTrue(view.isAttachedToWindow)
+        assertTrue(scope.isActive)
+
+        activity.windowManager.removeView(view)
+
+        assertFalse(view.isAttachedToWindow)
+        assertFalse(scope.isActive)
+
+        val latch = CountDownLatch(1)
+        var coroutineExecuted = false
+
+        scope.launch {
+            coroutineExecuted = true
+            latch.countDown()
+        }
+
+        assertFalse(latch.await(5, TimeUnit.SECONDS))
+        assertFalse(coroutineExecuted)
     }
 }